Fixed 4th and 5th Test
All 7 tests now pass. The problem with both of the tests was due to the fact that the endpoint only allows a 7 day range of data to be requested. So the problem with the SearchWithEndDateOnly was that it wasn't within 7 days in the past from the current day. The problem with the SearchWithValidDateRange test was that the date range was larger than 7 days. Because of this, the endpoint was sending back 400 responses.